chore(pull_out): move common utils
