+++ planning/behavior_path_planner/include/behavior_path_planner/util/pull_out/pull_out_parameters.hpp
@@ -13,8 +13,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OUT__PULL_OUT_PARAMETERS_HPP_
-#define BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OUT__PULL_OUT_PARAMETERS_HPP_
+#ifndef BEHAVIOR_PATH_PLANNER__UTIL__PULL_OUT__PULL_OUT_PARAMETERS_HPP_
+#define BEHAVIOR_PATH_PLANNER__UTIL__PULL_OUT__PULL_OUT_PARAMETERS_HPP_
 
 #include <string>
 #include <vector>
@@ -59,4 +59,4 @@ struct PullOutParameters
 
 }  // namespace behavior_path_planner
 
-#endif  // BEHAVIOR_PATH_PLANNER__SCENE_MODULE__PULL_OUT__PULL_OUT_PARAMETERS_HPP_
+#endif  // BEHAVIOR_PATH_PLANNER__UTIL__PULL_OUT__PULL_OUT_PARAMETERS_HPP_
